refactor(spectral tools): spectrogramcolors.h in internal
