+++ src/spectrogram/CMakeLists.txt
@@ -5,7 +5,6 @@
 declare_module(spectrogram)
 
 set(MODULE_SRC
-    ${CMAKE_CURRENT_LIST_DIR}/spectrogramcolors.h
     ${CMAKE_CURRENT_LIST_DIR}/ispectrogramconfiguration.h
     ${CMAKE_CURRENT_LIST_DIR}/ispectrogrampainter.h
     ${CMAKE_CURRENT_LIST_DIR}/spectrogrammodule.cpp
@@ -21,6 +20,7 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/internal/spectrogramcolorregister.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/spectrogramcolorregister.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/spectrogramcolors.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/spectrogramcolors.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/spectrogramconfiguration.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/spectrogramconfiguration.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/spectrogramutils.cpp
