+++ src/spectrogram/internal/spectrogramcolors.h
